Add 'fips-securitychecks' option and plumb this into the actual fips checks
