+++ providers/common/build.info
@@ -1,6 +1,7 @@
 SUBDIRS=der
 
 SOURCE[../libcommon.a]=provider_err.c provider_ctx.c
-$FIPSCOMMON=provider_util.c capabilities.c bio_prov.c digest_to_nid.c
-SOURCE[../libnonfips.a]=$FIPSCOMMON nid_to_name.c check_default.c
-SOURCE[../libfips.a]=$FIPSCOMMON check_fips.c
+$FIPSCOMMON=provider_util.c capabilities.c bio_prov.c digest_to_nid.c\
+            securitycheck.c
+SOURCE[../libnonfips.a]=$FIPSCOMMON nid_to_name.c securitycheck_default.c
+SOURCE[../libfips.a]=$FIPSCOMMON securitycheck_fips.c
