+++ providers/common/include/prov/securitycheck.h
@@ -7,7 +7,7 @@
  * https://www.openssl.org/source/license.html
  */
 
-/* Functions that have different implementations for the FIPS_MODULE */
+/* Functions that are common */
 int rsa_check_key(const RSA *rsa, int protect);
 int ec_check_key(const EC_KEY *ec, int protect);
 int dsa_check_key(const DSA *dsa, int sign);
@@ -15,8 +15,11 @@ int dh_check_key(const DH *dh);
 
 int digest_is_allowed(const EVP_MD *md);
 int digest_get_approved_nid_with_sha1(const EVP_MD *md, int sha1_allowed);
-int digest_rsa_sign_get_md_nid(const EVP_MD *md, ossl_unused int sha1_allowed);
 
 /* Functions that are common */
 int digest_md_to_nid(const EVP_MD *md, const OSSL_ITEM *it, size_t it_len);
 int digest_get_approved_nid(const EVP_MD *md);
+
+/* Functions that have different implementations for the FIPS_MODULE */
+int digest_rsa_sign_get_md_nid(const EVP_MD *md, int sha1_allowed);
+int securitycheck_enabled(void);
