+++ providers/implementations/signature/ecdsa.c
@@ -28,7 +28,7 @@
 #include "prov/providercommonerr.h"
 #include "prov/implementations.h"
 #include "prov/provider_ctx.h"
-#include "prov/check.h"
+#include "prov/securitycheck.h"
 #include "crypto/ec.h"
 #include "prov/der_ec.h"
